# Implement optimistic reader on the search phase