--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,2 @@
 .idea/
+vendor/