--- a/queryable/parquet_queryable.go
+++ b/queryable/parquet_queryable.go
@@ -240,7 +240,7 @@ func (b queryableShard) Query(ctx context.Context, sorted bool, mint, maxt int64
 	results := make([]prom_storage.ChunkSeries, 0, 1024)
 	rMtx := sync.Mutex{}
 
-	for i, group := range b.shard.LabelsFile().RowGroups() {
+	for rgi := range b.shard.LabelsFile().RowGroups() {
 		errGroup.Go(func() error {
 			cs, err := search.MatchersToConstraint(matchers...)
 			if err != nil {
@@ -250,7 +250,7 @@ func (b queryableShard) Query(ctx context.Context, sorted bool, mint, maxt int64
 			if err != nil {
 				return err
 			}
-			rr, err := search.Filter(ctx, group, cs...)
+			rr, err := search.Filter(ctx, b.shard.LabelsFile(), rgi, cs...)
 			if err != nil {
 				return err
 			}
@@ -259,7 +259,7 @@ func (b queryableShard) Query(ctx context.Context, sorted bool, mint, maxt int64
 				return nil
 			}
 
-			series, err := b.m.Materialize(ctx, i, mint, maxt, skipChunks, rr)
+			series, err := b.m.Materialize(ctx, rgi, mint, maxt, skipChunks, rr)
 			if err != nil {
 				return err
 			}
@@ -290,7 +290,7 @@ func (b queryableShard) LabelNames(ctx context.Context, limit int64, matchers []
 
 	results := make([][]string, len(b.shard.LabelsFile().RowGroups()))
 
-	for i, group := range b.shard.LabelsFile().RowGroups() {
+	for rgi := range b.shard.LabelsFile().RowGroups() {
 		errGroup.Go(func() error {
 			cs, err := search.MatchersToConstraint(matchers...)
 			if err != nil {
@@ -300,15 +300,15 @@ func (b queryableShard) LabelNames(ctx context.Context, limit int64, matchers []
 			if err != nil {
 				return err
 			}
-			rr, err := search.Filter(ctx, group, cs...)
+			rr, err := search.Filter(ctx, b.shard.LabelsFile(), rgi, cs...)
 			if err != nil {
 				return err
 			}
-			series, err := b.m.MaterializeLabelNames(ctx, i, rr)
+			series, err := b.m.MaterializeLabelNames(ctx, rgi, rr)
 			if err != nil {
 				return err
 			}
-			results[i] = series
+			results[rgi] = series
 			return nil
 		})
 	}
@@ -330,7 +330,7 @@ func (b queryableShard) LabelValues(ctx context.Context, name string, limit int6
 
 	results := make([][]string, len(b.shard.LabelsFile().RowGroups()))
 
-	for i, group := range b.shard.LabelsFile().RowGroups() {
+	for rgi, _ := range b.shard.LabelsFile().RowGroups() {
 		errGroup.Go(func() error {
 			cs, err := search.MatchersToConstraint(matchers...)
 			if err != nil {
@@ -340,15 +340,15 @@ func (b queryableShard) LabelValues(ctx context.Context, name string, limit int6
 			if err != nil {
 				return err
 			}
-			rr, err := search.Filter(ctx, group, cs...)
+			rr, err := search.Filter(ctx, b.shard.LabelsFile(), rgi, cs...)
 			if err != nil {
 				return err
 			}
-			series, err := b.m.MaterializeLabelValues(ctx, name, i, rr)
+			series, err := b.m.MaterializeLabelValues(ctx, name, rgi, rr)
 			if err != nil {
 				return err
 			}
-			results[i] = series
+			results[rgi] = series
 			return nil
 		})
 	}