--- a/queryable/parquet_queryable.go
+++ b/queryable/parquet_queryable.go
@@ -237,7 +237,7 @@ func (b queryableShard) Query(ctx context.Context, sorted bool, mint, maxt int64
 			if err != nil {
 				return err
 			}
-			err = search.Initialize(b.shard, cs...)
+			err = search.Initialize(b.shard.LabelsFile(), cs...)
 			if err != nil {
 				return err
 			}
@@ -287,7 +287,7 @@ func (b queryableShard) LabelNames(ctx context.Context, limit int64, matchers []
 			if err != nil {
 				return err
 			}
-			err = search.Initialize(b.shard, cs...)
+			err = search.Initialize(b.shard.LabelsFile(), cs...)
 			if err != nil {
 				return err
 			}
@@ -327,7 +327,7 @@ func (b queryableShard) LabelValues(ctx context.Context, name string, limit int6
 			if err != nil {
 				return err
 			}
-			err = search.Initialize(b.shard, cs...)
+			err = search.Initialize(b.shard.LabelsFile(), cs...)
 			if err != nil {
 				return err
 			}