Optimizes equality matchers and improves dictionary handling during materialization:

* Optimized equality matcher: The Equal constraint now leverages the optimized read path to fetch all relevant data pages in a single read. This reduces overmatching and improves read efficiency - similar to what is done during the materialization phase: https://github.com/prometheus-community/parquet-common/pull/48
* Materialization improvements: The dictionary page is now included as part of the optimistic read during materialization. This avoids an additional object store call when decoding dictionary-encoded columns.
    
 Results compared to mainline:
 
 ```
goos: darwin
goarch: arm64
pkg: github.com/prometheus-community/parquet-common/queryable
cpu: Apple M1 Pro
                                        │   /tmp/old   │              /tmp/new               │
                                        │    sec/op    │    sec/op     vs base               │
Select/SingleMetricAllSeries-10           398.4m ±  8%   392.1m ±  3%   -1.58% (p=0.041 n=6)
Select/SingleMetricReducedSeries-10       29.33m ±  9%   23.19m ±  6%  -20.92% (p=0.002 n=6)
Select/SingleMetricOneSeries-10           26.88m ±  3%   21.40m ±  3%  -20.38% (p=0.002 n=6)
Select/SingleMetricSparseSeries-10        54.26m ±  6%   46.45m ±  9%  -14.39% (p=0.002 n=6)
Select/NonExistentSeries-10               25.47m ±  3%   19.48m ±  5%  -23.51% (p=0.002 n=6)
Select/MultipleMetricsRange-10             1.859 ±  4%    1.791 ±  3%   -3.64% (p=0.041 n=6)
Select/MultipleMetricsSparse-10           553.8m ±  8%   539.0m ±  6%        ~ (p=0.310 n=6)
Select/NegativeRegexSingleMetric-10       383.9m ±  7%   369.2m ±  5%        ~ (p=0.310 n=6)
Select/NegativeRegexMultipleMetrics-10     1.103 ±  3%    1.081 ±  7%        ~ (p=0.310 n=6)
Select/ExpensiveRegexSingleMetric-10      67.42m ± 12%   61.43m ± 16%        ~ (p=0.093 n=6)
Select/ExpensiveRegexMultipleMetrics-10   185.5m ±  3%   184.5m ±  4%        ~ (p=0.485 n=6)
geomean                                   164.4m         148.4m         -9.71%

                                        │      /tmp/old      │                  /tmp/new                   │
                                        │ bytes_get_range/op │ bytes_get_range/op  vs base                 │
Select/SingleMetricAllSeries-10                  9.137M ± 0%          5.774M ± 0%  -36.80% (p=0.002 n=6)
Select/SingleMetricReducedSeries-10             2029.3k ± 0%          423.4k ± 0%  -79.13% (p=0.002 n=6)
Select/SingleMetricOneSeries-10                 1978.0k ± 0%          376.5k ± 0%  -80.97% (p=0.002 n=6)
Select/SingleMetricSparseSeries-10               5.487M ± 0%          2.072M ± 0%  -62.25% (p=0.002 n=6)
Select/NonExistentSeries-10                     1938.7k ± 0%          366.6k ± 0%  -81.09% (p=0.002 n=6)
Select/MultipleMetricsRange-10                   38.01M ± 0%          30.84M ± 0%  -18.85% (p=0.002 n=6)
Select/MultipleMetricsSparse-10                  16.41M ± 0%          14.62M ± 0%  -10.91% (p=0.002 n=6)
Select/NegativeRegexSingleMetric-10              9.394M ± 0%          6.426M ± 0%  -31.60% (p=0.002 n=6)
Select/NegativeRegexMultipleMetrics-10           31.58M ± 0%          27.39M ± 0%  -13.27% (p=0.002 n=6)
Select/ExpensiveRegexSingleMetric-10             3.924M ± 0%          2.316M ± 0%  -40.97% (p=0.002 n=6)
Select/ExpensiveRegexMultipleMetrics-10          14.74M ± 0%          14.74M ± 0%        ~ (p=1.000 n=6) ¹
geomean                                          7.492M               3.721M       -50.33%
¹ all samples are equal

                                        │   /tmp/old   │              /tmp/new              │
                                        │    get/op    │   get/op    vs base                │
Select/SingleMetricAllSeries-10           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/SingleMetricReducedSeries-10       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/SingleMetricOneSeries-10           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/SingleMetricSparseSeries-10        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/NonExistentSeries-10               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/MultipleMetricsRange-10            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/MultipleMetricsSparse-10           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/NegativeRegexSingleMetric-10       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/NegativeRegexMultipleMetrics-10    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/ExpensiveRegexSingleMetric-10      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/ExpensiveRegexMultipleMetrics-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                              ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                        │   /tmp/old   │               /tmp/new                │
                                        │ get_range/op │ get_range/op  vs base                 │
Select/SingleMetricAllSeries-10           19.163k ± 0%    8.400k ± 0%  -56.17% (p=0.002 n=6)
Select/SingleMetricReducedSeries-10       10.925k ± 0%    3.654k ± 0%  -66.55% (p=0.002 n=6)
Select/SingleMetricOneSeries-10           10.905k ± 0%    3.624k ± 0%  -66.77% (p=0.002 n=6)
Select/SingleMetricSparseSeries-10        18.001k ± 0%    6.000k ± 0%  -66.67% (p=0.002 n=6)
Select/NonExistentSeries-10               10.763k ± 0%    3.600k ± 0%  -66.55% (p=0.002 n=6)
Select/MultipleMetricsRange-10             72.41k ± 0%    58.01k ± 0%  -19.89% (p=0.002 n=6)
Select/MultipleMetricsSparse-10            47.21k ± 0%    43.61k ± 0%   -7.63% (p=0.002 n=6)
Select/NegativeRegexSingleMetric-10        25.08k ± 0%    15.11k ± 0%  -39.76% (p=0.002 n=6)
Select/NegativeRegexMultipleMetrics-10     81.76k ± 0%    73.33k ± 0%  -10.30% (p=0.002 n=6)
Select/ExpensiveRegexSingleMetric-10       18.69k ± 0%    11.46k ± 0%  -38.70% (p=0.002 n=6)
Select/ExpensiveRegexMultipleMetrics-10    62.10k ± 0%    62.10k ± 0%        ~ (p=1.000 n=6) ¹
geomean                                    25.95k         14.22k       -45.23%
¹ all samples are equal

                                        │   /tmp/old    │              /tmp/new               │
                                        │   series/op   │  series/op   vs base                │
Select/SingleMetricAllSeries-10           300.0k ± 0%     300.0k ± 0%       ~ (p=1.000 n=6) ¹
Select/SingleMetricReducedSeries-10       3.000k ± 0%     3.000k ± 0%       ~ (p=1.000 n=6) ¹
Select/SingleMetricOneSeries-10            1.000 ± 0%      1.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/SingleMetricSparseSeries-10        5.000k ± 0%     5.000k ± 0%       ~ (p=1.000 n=6) ¹
Select/NonExistentSeries-10                0.000 ± 0%      0.000 ± 0%       ~ (p=1.000 n=6) ¹
Select/MultipleMetricsRange-10            1.200M ± 0%     1.200M ± 0%       ~ (p=1.000 n=6) ¹
Select/MultipleMetricsSparse-10           300.0k ± 0%     300.0k ± 0%       ~ (p=1.000 n=6) ¹
Select/NegativeRegexSingleMetric-10       234.0k ± 0%     234.0k ± 0%       ~ (p=1.000 n=6) ¹
Select/NegativeRegexMultipleMetrics-10    702.0k ± 0%     702.0k ± 0%       ~ (p=1.000 n=6) ¹
Select/ExpensiveRegexSingleMetric-10      6.000k ± 0%     6.000k ± 0%       ~ (p=1.000 n=6) ¹
Select/ExpensiveRegexMultipleMetrics-10    0.000 ± 0%      0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                               ²                +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                        │   /tmp/old   │              /tmp/new              │
                                        │     B/op     │     B/op      vs base              │
Select/SingleMetricAllSeries-10           777.0Mi ± 0%   775.3Mi ± 0%  -0.21% (p=0.002 n=6)
Select/SingleMetricReducedSeries-10       71.67Mi ± 0%   71.03Mi ± 0%  -0.89% (p=0.002 n=6)
Select/SingleMetricOneSeries-10           66.98Mi ± 0%   66.35Mi ± 0%  -0.94% (p=0.002 n=6)
Select/SingleMetricSparseSeries-10        131.1Mi ± 0%   130.3Mi ± 0%  -0.61% (p=0.002 n=6)
Select/NonExistentSeries-10               64.08Mi ± 0%   63.45Mi ± 0%  -0.98% (p=0.002 n=6)
Select/MultipleMetricsRange-10            3.108Gi ± 0%   3.104Gi ± 0%  -0.14% (p=0.002 n=6)
Select/MultipleMetricsSparse-10           1.033Gi ± 0%   1.032Gi ± 0%  -0.11% (p=0.002 n=6)
Select/NegativeRegexSingleMetric-10       716.1Mi ± 0%   714.7Mi ± 0%  -0.19% (p=0.002 n=6)
Select/NegativeRegexMultipleMetrics-10    2.170Gi ± 0%   2.167Gi ± 0%  -0.12% (p=0.002 n=6)
Select/ExpensiveRegexSingleMetric-10      192.7Mi ± 0%   192.2Mi ± 0%  -0.27% (p=0.002 n=6)
Select/ExpensiveRegexMultipleMetrics-10   539.8Mi ± 0%   539.8Mi ± 0%  -0.01% (p=0.002 n=6)
geomean                                   369.4Mi        367.9Mi       -0.41%

                                        │  /tmp/old   │              /tmp/new              │
                                        │  allocs/op  │  allocs/op   vs base               │
Select/SingleMetricAllSeries-10           9.845M ± 0%   9.780M ± 0%   -0.66% (p=0.002 n=6)
Select/SingleMetricReducedSeries-10       289.6k ± 0%   260.3k ± 0%  -10.11% (p=0.002 n=6)
Select/SingleMetricOneSeries-10           266.2k ± 0%   237.2k ± 0%  -10.93% (p=0.002 n=6)
Select/SingleMetricSparseSeries-10        647.9k ± 0%   597.8k ± 0%   -7.73% (p=0.002 n=6)
Select/NonExistentSeries-10               191.6k ± 0%   162.9k ± 0%  -14.96% (p=0.002 n=6)
Select/MultipleMetricsRange-10            39.71M ± 0%   39.56M ± 0%   -0.39% (p=0.002 n=6)
Select/MultipleMetricsSparse-10           10.82M ± 0%   10.78M ± 0%   -0.35% (p=0.002 n=6)
Select/NegativeRegexSingleMetric-10       8.353M ± 0%   8.296M ± 0%   -0.68% (p=0.002 n=6)
Select/NegativeRegexMultipleMetrics-10    24.50M ± 0%   24.41M ± 0%   -0.37% (p=0.002 n=6)
Select/ExpensiveRegexSingleMetric-10      915.6k ± 0%   888.0k ± 0%   -3.02% (p=0.002 n=6)
Select/ExpensiveRegexMultipleMetrics-10   1.796M ± 0%   1.796M ± 0%        ~ (p=0.108 n=6)
geomean                                   2.376M        2.266M        -4.62%
 ```