@alanprot 2025-07-02
I followed up about my previous comment.

Now we are are not using the gap partitioner anymore and we are only adding the dictionary page on the "single read" if the gap is < than the `PagePartitioningMaxGapSize`

@MichaHoffmann WDYT?